% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to dataset}
\usage{
add_metadata(dat, metadat)
}
\arguments{
\item{dat}{A dataframe or tibble of data. A dataframe.}

\item{metadat}{A dataframe or tibble of metadata. A dataframe.}
}
\value{
A tibble.
}
\description{
Given a dataset with observations as rows and variables as columns,
and given a dataset of metadata with variables as rows and
information as columns, add metadata to the variables present in
the dataset.
}
