% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edcs.R
\name{adjust_chemicals_for_creatinine}
\alias{adjust_chemicals_for_creatinine}
\title{Title}
\usage{
adjust_chemicals_for_creatinine(dat, exposures, covariates, creatinine, method)
}
\arguments{
\item{method}{}
}
\description{
Title
}
