% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{bound_outcome}
\alias{bound_outcome}
\title{Bound a outcome variable}
\usage{
bound_outcome(dat, var)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A tibble.}

\item{var}{The variable name corresponding to the outcome to be bounded. A string.}
}
\value{
A dataframe containing the bounded outcome. A tibble.
}
\description{
In order to be able to use e.g., TMLE, with a continuous outcome,
it is necessary to bound it between 0 and 1. This function performs
the necessary steps.
}
