% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_model}
\alias{check_model}
\title{Tidy checks for model fit}
\usage{
check_model(model, path_save_res)
}
\arguments{
\item{model}{A fitted model. An object of class corresponding to the chosen model.}

\item{path_save_res}{A optional path to directory to save figures. A string.}
}
\description{
Tidy checks for model fit
}
