% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{convert_time_season}
\alias{convert_time_season}
\title{Convert time variables to season}
\usage{
convert_time_season(dat, cols)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A dataframe.}

\item{cols}{A vector of column names as strings. A vector.}
}
\value{
A dataframe with additional columns for seasons. A dataframe.
}
\description{
Convert time variables to season
}
