% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_data}
\alias{create_data}
\title{Create toy dataset given options}
\usage{
create_data(variables, num_rows, num_cols, ids)
}
\arguments{
\item{variables}{.}

\item{num_rows}{.}

\item{num_cols}{.}

\item{ids}{.}
}
\value{
.
}
\description{
Given a list of options, creates a toy dataset satisfying
those requirements.
}
