% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake.R
\name{create_fake_data}
\alias{create_fake_data}
\title{Create fake dataset based on actual one}
\usage{
create_fake_data(dat)
}
\arguments{
\item{dat}{A dataset with variables as columns. A data.table.}
}
\value{
A random dataset with matching variables. A data.table.
}
\description{
Given a dataset with variables as columns, create a
random dataset that can be stored on the local machine
without security and privacy issues. The type of the
variables is matched.
}
