% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_mapping_labels}
\alias{create_mapping_labels}
\title{Create a variable's dictionary for its labels}
\usage{
create_mapping_labels(labels, codes)
}
\arguments{
\item{labels}{A string of labels separated by a comma and a space. A string.}

\item{code}{A string of values separated by a comma. A string.}
}
\value{
A named list mapping labels to values.
}
\description{
Given the labels and the codes of a variable,
create a named list.
}
