% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{describe_data}
\alias{describe_data}
\title{A function to describe the used dataframe}
\usage{
describe_data(dat, id_var, by_var)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A dataframe.}

\item{id_var}{The variable name to be used to identify subjects. A string.}

\item{by_var}{The variable name to group by. A string.}
}
\value{
A nested named list containing the objects corresponding to the
steps described above.
}
\description{
Given a dataframe, this function performs the following steps:
\itemize{
\item Simple diagnose of numerical and categorical
variables with the \link[dlookr]{diagnose_numeric} and
\link[dlookr]{diagnose_category} functions, respectively.
\item Diagnose of numerical and categorical
variables with the \link[dlookr]{diagnose_numeric} and
\link[dlookr]{diagnose_category} functions, respectively,
after grouping by a factor (e.g., cohort).
\item Concise summary of continuous and categorical variables with
the \link[gtsummary]{tbl_summary} function.
\item Detailed summary of continuous and categorical variables with
the \link[gtsummary]{tbl_summary} function.
\item Visualization of the correlation structure of the dataframe using
the \link[corrr]{correlate} and \link[corrr]{rplot} functions,
with both Pearson's and Spearman's correlation.
}
}
