% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{df_mapping_vars}
\alias{df_mapping_vars}
\title{Create dataframe for mapping of variables}
\usage{
df_mapping_vars(dictionary, description)
}
\arguments{
\item{dictionary}{A named list of vectors. A list.}

\item{description}{A dataframe.}
}
\value{
A dataframe.
}
\description{
Given a dictionary of the form \code{var_dag = c()}, and an optional
dataframe of variables' description, creates a unnested
dataframe of the form \verb{var_dag, var_helix, description}.
}
