% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edcs.R
\name{edcs_information}
\alias{edcs_information}
\title{Return information on chemicals of interest}
\usage{
edcs_information()
}
\value{
A \link[data.table]{data.table} object. Each entry has
information on full and short names, chemical class, SMILES, PubChem, CTD,
and ExposomeExplorer IDs, eventual parental compound.
}
\description{
Return information on chemicals of interest
}
