% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{estimate_marginal_effects}
\alias{estimate_marginal_effects}
\title{Title}
\usage{
estimate_marginal_effects(dat, outcome, exposure, covariates, model)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A tibble.}

\item{exposure}{The name of the variable corresponding to the exposure. A string.}

\item{covariates}{A vector of covariates' names. A vector.}

\item{model}{}
}
\description{
Title
}
