% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{estimate_selection_weights}
\alias{estimate_selection_weights}
\title{Title}
\usage{
estimate_selection_weights(
  dat,
  idxs_selected,
  id_var,
  filter_out,
  list_covars,
  method,
  method_args,
  trim_weights,
  threshold_trim
)
}
\arguments{
\item{threshold_trim}{}
}
\description{
Title
}
