% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{find_intersection_confounders}
\alias{find_intersection_confounders}
\title{Find intersection of confounders}
\usage{
find_intersection_confounders(dag, type_mas, type_effect)
}
\arguments{
\item{dag}{A \pkg{dagitty} object.}

\item{type_mas}{The type of adjustment set.}

\item{type_effect}{The type of effect to be estimated.}
}
\value{
A list.
}
\description{
Given a DAG with multiple adjustment sets, find the
intersection of all the nodes.
}
