% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{from_dagitty_to_ggdag}
\alias{from_dagitty_to_ggdag}
\title{Convert a DAG to a formatted string}
\usage{
from_dagitty_to_ggdag(dag)
}
\arguments{
\item{dag}{A textual description of a DAG that can
be read by the \link[dagitty]{dagitty} function. A string.}
}
\value{
A formatted string.
}
\description{
Convert a DAG to a formatted
string that can be used to create DAGs with \link[ggdag]{ggdag}
using its formula-based input method.
}
