% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{handle_loq}
\alias{handle_loq}
\title{Handle values below the limit of quantification}
\usage{
handle_loq(dat, strategy)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A tibble.}

\item{strategy}{}
}
\description{
Handle values below the limit of quantification
}
