% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{handle_standardization}
\alias{handle_standardization}
\title{Title}
\usage{
handle_standardization(dat, id_var, by_var, center_fun, scale_fun)
}
\arguments{
\item{scale_fun}{}
}
\description{
Title
}
