% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{handle_transformation}
\alias{handle_transformation}
\title{Title}
\usage{
handle_transformation(dat, id_var, by_var, transformation_fun)
}
\arguments{
\item{transformation_fun}{}
}
\description{
Title
}
