% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{load_ctd}
\alias{load_ctd}
\title{Load local CTD file}
\usage{
load_ctd(path, filter_evidence)
}
\arguments{
\item{path}{Path to the CTD file. A string.}

\item{filter_evidence}{Indicates whether to filter the type of
direct evidence. A bool.}
}
\value{
A dataframe containing the CTD database.
}
\description{
Load local CTD file
}
