% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{map_covariates}
\alias{map_covariates}
\title{Mapping from DAG nodes to variables}
\usage{
map_covariates(adj_set, params)
}
\arguments{
\item{adj_set}{.}

\item{params}{.}
}
\value{
.
}
\description{
Maps the names of the nodes in a DAG to the corresponding
variables in the dataset.
}
