% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{map_covars}
\alias{map_covars}
\title{Mapping between causal nodes and variables' names}
\usage{
map_covars()
}
\arguments{
\item{A.}{.}
}
\value{
A list of variables' names.
}
\description{
Mapping between causal nodes and variables' names
}
