% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{minimize_missings}
\alias{minimize_missings}
\title{Select confounders to minimize missing values}
\usage{
minimize_missings(dat, meta, adjustment_sets, by_var)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A dataframe.}

\item{meta}{A dataframe containing the metadata. A dataframe.}

\item{adjustment_sets}{A list of adjustment sets. A character vector.}

\item{by_var}{The variable name to group by. A string.}
}
\value{
An integer corresponding to the index of the adjustment
set that minimizes the number of missing values.
}
\description{
Given a dataset and a list of adjustment sets
identified with the \link[dagitty]{adjustmentSets} function,
find the set that minimizes the number of missing values
for the confounders.
}
