% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myphd-package.R
\docType{package}
\name{myphd-package}
\alias{myphd}
\alias{myphd-package}
\title{myphd: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lorenzofabbri.github.io/myphd/}
}

}
\keyword{internal}
