% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{plot_ctd}
\alias{plot_ctd}
\title{Visualize the CTD dataset after loading}
\usage{
plot_ctd(dat, group)
}
\arguments{
\item{dat}{The CTD dataset. A dataframe.}

\item{group}{The faceting variable. A string.}
}
\value{
A ggplot object.
}
\description{
Given a dataframe with the data from the CTD, create a bar plot of the
categories.
}
