% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{test.npsem}
\alias{test.npsem}
\title{Test conditional independencies.
\code{test.npsem} checks whether the assumptions
encoded in the DAG are consistent with the data.}
\usage{
test.npsem(dag.code, dat, params)
}
\arguments{
\item{dag.code}{Input DAG name. A string.}

\item{dat}{The dataset to test the DAG against. A dataframe.}

\item{params}{A named list containing eventual parameters.}
}
\value{
The results of \code{dagitty::localTests}.
}
\description{
Test conditional independencies.
\code{test.npsem} checks whether the assumptions
encoded in the DAG are consistent with the data.
}
