% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{visualize.dag}
\alias{visualize.dag}
\title{Visualize DAG.
\code{visualize.dag} is a wrapper function to plot DAGs
using \pkg{ggdag}.}
\usage{
visualize.dag(dag.code)
}
\arguments{
\item{dag.code}{Input DAG name. A string.}
}
\value{
A \pkg{ggplot} object.
}
\description{
Visualize DAG.
\code{visualize.dag} is a wrapper function to plot DAGs
using \pkg{ggdag}.
}
