% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{add_metadata}
\alias{add_metadata}
\title{Add metadata to dataframe}
\usage{
add_metadata(dat, metadat, categorical_types, cols_to_exclude)
}
\arguments{
\item{dat}{A dataframe or tibble of data. A dataframe.}

\item{metadat}{A dataframe or tibble of metadata.
Currently, the following fields are required:
\itemize{
\item \code{description}, a text description of the variable.
\item \code{remark}, any remark about the variable (e.g., if certain values were imputed).
\item \code{type}, type of variable (e.g., numerical).
\item \code{comments}, any other remark about the variable (e.g., unit of measurement).
\item \code{dag}, name of the variable as present in the DAG.
\item \code{period}, period of measurement of the variable (e.g., pregnancy).
\item \code{label}, labels of the levels in case of categorical variables,
separated by a comma and a space.
\item \code{code}, levels of the variable if categorical, separated by a comma.
A dataframe.
}}

\item{categorical_types}{Identifiers, as strings, of categorical types. A vector.}
}
\value{
A dataframe containing annotated variables.
}
\description{
Given a dataframe with observations as rows and variables as columns,
and given a dataframe of metadata with variables as rows and
information as columns, add metadata to the variables present in
the dataframe.
}
