% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data.R
\name{create_df_marginal_comparisons}
\alias{create_df_marginal_comparisons}
\title{Title}
\usage{
create_df_marginal_comparisons(dat, var, percentiles, by_var)
}
\arguments{
\item{by_var}{}
}
\description{
Title
}
