% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{create_formula}
\alias{create_formula}
\title{Title}
\usage{
create_formula(
  dat,
  outcome,
  exposure,
  covariates,
  method,
  add_inter_exposure,
  add_splines_exposure,
  df_splines,
  threshold_smooth,
  threshold_k
)
}
\arguments{
\item{df_splines}{}
}
\description{
Title
}
