% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edcs.R
\name{edcs_information}
\alias{edcs_information}
\title{Return information on non-persistent EDCs}
\usage{
edcs_information()
}
\value{
A data.table object.
}
\description{
Return information on non-persistent EDCs
}
