% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_missings}
\alias{explore_missings}
\title{A function to explore the missing values of a dataset}
\usage{
explore_missings(dat, id_var, by_var, path_save)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A tibble.}

\item{id_var}{The variable name to be used to identify subjects. A string.}

\item{by_var}{The variable name to group by. A string.}

\item{path_save}{The path where to store the missingness report. A string.}
}
\value{
A named list containing the objects corresponding to the
steps described above.
}
\description{
Given a dataset, this function performs the following steps:
\itemize{
\item Summary of missing values for the variables, grouped by a factor,
using the \link[naniar]{miss_var_summary} function.
The information is the same as that provided in the previous step,
but also in long format.
\item Summary of missing values for the cases
using the \link[naniar]{miss_case_summary} function.
\item Missingness report using the \link[visdat]{vis_miss} function.
\item Little's test statistic to assess if data is missing
completely at random (MCAR), using the \link[naniar]{mcar_test} function.
The null hypothesis in this test is that the data is MCAR.
}
}
