% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_data.R
\name{explore_shift}
\alias{explore_shift}
\title{Title}
\usage{
explore_shift(
  dat,
  shift_type,
  shift_amount,
  shift_lower_bound,
  shift_upper_bound
)
}
\arguments{
\item{shift_func}{}
}
\description{
Title
}
