% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{extract_cohort}
\alias{extract_cohort}
\title{Extract the cohort label from the subject ID}
\usage{
extract_cohort(dat, id_var, st, en)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A dataframe.}

\item{id_var}{The variable name to be used to identify subjects. A string.}

\item{st}{Integer indicating the first element to be extracted. An integer.}

\item{en}{Integer indicating the last element to be extracted. An integer.}
}
\value{
A dataframe containing a new column named \code{cohort}. A dataframe.
}
\description{
This function extracts the cohort ID (e.g., SAB) from the subject ID
(e.g., SubjectID). By default it takes the first three characters.
}
