% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{find_common_confounders}
\alias{find_common_confounders}
\title{Find common confounders}
\usage{
find_common_confounders(dag, type_mas, type_effect)
}
\arguments{
\item{dag}{A \link[dagitty]{dagitty} object.}

\item{type_mas}{The type of adjustment set. A string.}

\item{type_effect}{The type of effect to be estimated. A string.}
}
\value{
The list of common nodes. A list.
}
\description{
Given a DAG with multiple adjustment sets, find
all the common nodes.
}
