% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighting.R
\name{fit_model_weighted}
\alias{fit_model_weighted}
\title{Title}
\usage{
fit_model_weighted(dat, exposure, covariates, weights, method, method_args)
}
\arguments{
\item{method_args}{}
}
\description{
Title
}
