% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{fit_model_weighted}
\alias{fit_model_weighted}
\title{Fit various models with weighting}
\usage{
fit_model_weighted(
  dat,
  outcome,
  exposure,
  covariates,
  weights,
  method,
  method_args
)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A tibble.}

\item{exposure}{The name of the variable corresponding to the exposure. A string.}

\item{covariates}{A vector of covariates' names. A vector.}

\item{weights}{The \code{weights} element of the result of the call
to \code{\link[=estimate_weights]{estimate_weights()}}. A \link[WeightIt]{weightit} object.}

\item{method_args}{A named list with the following variables:
\itemize{
\item \code{family}, .
\item \code{add_inter_exposure}, .
\item \code{add_splines_exposure}, .
\item \code{df_splines}, .
\item \code{threshold_smooth}, .
\item \code{threshold_k}, .
}}
}
\description{
Fit various models with weighting
}
