% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{from_dagitty_to_ggdag}
\alias{from_dagitty_to_ggdag}
\title{Convert a \pkg{dagitty} DAG to a formatted string}
\usage{
from_dagitty_to_ggdag(dag)
}
\arguments{
\item{dag}{A \pkg{dagitty} DAG. A string.}
}
\value{
A formatted string.
}
\description{
Convert a DAG from \pkg{dagitty} to a formatted
string that can be used to create DAGs with \pkg{ggdag}
using its formula-based input method.
}
