% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{handle_creatinine_confounding}
\alias{handle_creatinine_confounding}
\title{Title}
\usage{
handle_creatinine_confounding(
  dat,
  covariates,
  id_var,
  covariates_names,
  creatinine,
  method,
  method_fit_args
)
}
\arguments{
\item{method_fit_args}{}
}
\description{
Title
}
