% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{handle_llodq}
\alias{handle_llodq}
\title{Various strategies to handle values below the limit of detection/quantification}
\usage{
handle_llodq(dat, method)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A tibble.}

\item{method}{}
}
\description{
Various strategies to handle values below the limit of detection/quantification
}
