% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myphd-package.R
\docType{package}
\name{myphd-package}
\alias{myphd}
\alias{myphd-package}
\title{myphd: A Small R Package for my PhD}
\description{
An R package written during my PhD years to automate certain tasks in epidemiology and causa inference reseaerch.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lorenzofabbri.github.io/myphd/}
  \item \url{https://isglobal-cep.github.io/myphd/}
}

}
\author{
\strong{Maintainer}: Lorenzo Fabbri \email{lorenzo.fabbri@isglobal.org} (\href{https://orcid.org/0000-0003-3031-322X}{ORCID})

}
\keyword{internal}
