% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{preproc_data}
\alias{preproc_data}
\title{Basic pre-processing of datasets}
\usage{
preproc_data(
  dat,
  dat_desc = NULL,
  covariates,
  outcome,
  dat_llodq = NULL,
  dic_steps,
  id_var,
  by_var
)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A tibble.}

\item{covariates}{A dataframe containing additional variables. A tibble.}

\item{outcome}{A string indicating the outcome variable. A string.}

\item{dic_steps}{A nested named list of steps to perform. A list. It can
include the following elements:
\itemize{
\item \code{llodq}, to handle values <LOD/LOQ. A named list with elements:
\item \code{id_val}, .
\item \code{method}, the method to be used. A string.
\item \code{creatinine_threshold}, .
\item \code{threshold_within}, .
\item \code{threshold_overall}, .
\item \code{tune_sigma}, .
\item \code{missings}, to handle missing values. A named list with elements:
\item \code{threshold_within}, the missing value threshold within each group. An integer.
\item \code{threshold_overall}, the overall missing value threshold. An integer.
\item \code{selected_covariates}, .
\item \code{method_imputation}, .
\item \code{creatinine}, to handle confounding by dilution. A named list with elements:
\item \code{method}, .
\item \code{method_fit_args}, options for fitting the models. A list.
\item \code{creatinine_covariates_names}, .
\item \code{creatinine_name}, .
\item \code{standardization}, to standardize variables. A named list with elements:
\item \code{center_fun}, the centering function (e.g., \code{median}).
\item \code{scale_fun}, the scaling function (e.g., \code{IQR}).
\item \code{bound}, to bound the outcome variable. A named list with elements:
}}

\item{id_var}{The variable name to be used to identify subjects. A string.}

\item{by_var}{The variable name to group by. A string.}
}
\value{
A pre-processed dataset. A tibble.
}
\description{
This function performs basic pre-processing of a given dataset:
\itemize{
\item Data cleaning:
\itemize{
\item Removal of incorrect/unwanted variables.
\item Remove of incorrect/unwanted subjects.
\item Handling of missing values.
}
\item Variable transformations:
\itemize{
\item Transformation (e.g., log-transformation, bound outcome).
\item Standardization (zero mean and unit standard deviation).
\item Normalization (subtract the minimum and divide by the range).
}
}
}
