% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{preproc_data}
\alias{preproc_data}
\title{Basic pre-processing of dataframes}
\usage{
preproc_data(
  dat,
  dat_desc = NULL,
  covariates,
  outcome,
  dat_llodq = NULL,
  dic_steps,
  id_var,
  by_var
)
}
\arguments{
\item{dat}{A dataframe containing the variables of interest. A dataframe.}

\item{dat_desc}{A optional dataframe containing the same variables but
with information on the type of values (e.g., 2 for value <LOD).
This is necessary to distinguish the type of missing values
(e.g., missing because <LOD or because no sample was available). A dataframe.}

\item{covariates}{A dataframe containing additional variables. A dataframe.}

\item{outcome}{A string indicating the outcome variable. A string.}

\item{dat_llodq}{A optional dataframe containing the LOD/LOQ values (\code{val})
for the variables (\code{var}) of interest. A dataframe.}

\item{dic_steps}{A ordered, nested named list of steps to perform. A list. It can
include the following elements:
\itemize{
\item \code{llodq}, to handle values <LOD/LOQ. A named list with elements:
\item \code{id_val}, which values in \code{dat_desc} should be considered. A vector of integers.
\item \code{method}, the method to be used. Currently, only a replacement
approach is supported. A string.
\item \code{creatinine_threshold}, subjects for which the creatinine levels are
below this value will not be processed. Currently not used. A double.
\item \code{threshold_within}, the threshold within each group for the
fraction of values corresponding to \code{id_val}. An integer.
\item \code{threshold_overall}, the overall threshold for the
fraction of values corresponding to \code{id_val}. An integer.
\item \code{tune_sigma}, currently not supported. A double.
\item \code{missings}, to handle missing values. A named list with elements:
\item \code{threshold_within}, the missing value threshold within each group. An integer.
\item \code{threshold_overall}, the overall missing value threshold. An integer.
\item \code{use_additional_covariates}, .
\item \code{selected_covariates}, a vector of covariates' names. A vector.
\item \code{method_imputation}, method to be used to impute values. A string.
\item \code{k}, the number of nearest neighbors to use for kNN. An integer.
\item \code{path_save_res}, path to directory where to save figures.
Currently, variables can be imputed in a univariate way (\code{univariate}), using
selected covariates (\code{selected}), or all the covariates
available in \code{covariates} (\code{all}). A string.
\item \code{creatinine}, to handle confounding by dilution. A named list with elements:
\item \code{method}, the method to be used. Currently, only covariate-adjusted
standardization is implemented (\code{cas}). A string.
\item \code{method_fit_args}, options for fitting the models.
Currently, only the family to be used within \link[stats]{glm}. A list.
\item \code{creatinine_covariates_names}, .
\item \code{creatinine_name}, .
\item \code{path_save_res}, path to directory where to save figures.
\item \code{transform}, to transform variables. A named list with elements:
\item \code{transformation_fun}, the transformation function (e.g., \code{log}).
\item \code{standardization}, to standardize variables. A named list with elements:
\item \code{center_fun}, the centering function (e.g., \code{median}).
\item \code{scale_fun}, the scaling function (e.g., \code{IQR}).
\item \code{bound}, to bound the outcome variable. A named list with elements:
}}

\item{id_var}{The variable name to be used to identify subjects. A string.}

\item{by_var}{The variable name to group by. A string.}
}
\value{
A pre-processed dataframe. A dataframe.
}
\description{
This function performs basic pre-processing of a given dataframe:
\itemize{
\item Data cleaning:
\itemize{
\item Handling values below the limit of detection (LOD) or quantification (LOQ).
\item Handling missing values.
}
\item Variable transformations:
\itemize{
\item Standardization (with custom center and scale functions).
\item Correcting for urine dilution with creatinine.
\item Outcome bounding (for e.g., TMLE).
}
}
}
