% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{test_npsem}
\alias{test_npsem}
\title{Test conditional independencies}
\usage{
test_npsem(dag_code, dat, params)
}
\arguments{
\item{dag_code}{Input DAG name. A string.}

\item{dat}{The dataset to test the DAG against. A dataframe.}

\item{params}{A named list containing eventual parameters.}
}
\value{
The results of \code{dagitty::localTests}.
}
\description{
Check whether the assumptions encoded in the DAG
are consistent with the data.
}
