% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dags.R
\name{visualize_dag}
\alias{visualize_dag}
\title{Visualize DAG}
\usage{
visualize_dag(dag)
}
\arguments{
\item{dag}{A \pkg{dagitty} object.}
}
\value{
A \pkg{ggdag} object.
}
\description{
Wrapper function to plot DAGs using \pkg{ggdag}.
}
